requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioredis==2.0.1
aiosignal==1.4.0
aiosmtplib==4.0.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
backports.asyncio.runner==1.2.0
backrefs==5.9
bcrypt==4.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.1
black==25.1.0
bleach==6.2.0
cachetools==6.1.0
celery==5.5.3
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
coverage==7.9.2
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
emails==0.6
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
factory_boy==3.3.3
Faker==37.4.2
fastapi==0.115.5
fastapi-cache2==0.2.2
fastapi-limiter==0.1.6
fastapi-users==14.0.1
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.7.0
ghp-import==2.1.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
identify==2.6.12
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kombu==5.5.4
langdetect==1.0.9
limits==5.4.0
lxml==6.0.0
makefun==1.16.0
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib-inline==0.1.7
memory-profiler==0.61.0
mergedeep==1.3.4
mistune==3.1.3
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
more-itertools==10.7.0
motor==3.6.0
moviepy==2.2.1
multidict==6.6.3
mypy==1.17.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
nodeenv==1.9.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.0
overrides==7.7.0
packaging==24.2
paginate==0.5.7
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
phonenumbers==9.0.9
pillow==11.3.0
platformdirs==4.3.6
pluggy==1.6.0
pre_commit==4.2.0
premailer==3.10.0
proglog==0.1.12
prometheus_client==0.22.1
prompt_toolkit==3.0.50
propcache==0.3.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.16
pymongo==4.9.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-jose==3.3.0
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==26.2.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.26.0
rsa==4.9
ruff==0.12.3
Send2Trash==1.8.3
sentry-sdk==2.33.0
simple-websocket==1.1.0
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.41.3
structlog==25.4.0
terminado==0.18.1
textblob==0.19.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
validators==0.35.0
vine==5.1.0
virtualenv==20.31.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webdavclient3==3.14.6
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
